Migrated todoListGenerator function into its own script and imported back into domManip to clean up file
